Rework a lot of Sodium compatibility code as a bodge

+++ src/main/java/net/coderbot/iris/shadows/frustum/BoxCuller.java
@@ -30,7 +30,7 @@ public class BoxCuller {
 				(float) aabb.maxX, (float) aabb.maxY, (float) aabb.maxZ);
 	}
 
-	public boolean isCulled(float minX, float minY, float minZ, float maxX, float maxY, float maxZ) {
+	public boolean isCulled(double minX, double minY, double minZ, double maxX, double maxY, double maxZ) {
 		if (maxX < this.minAllowedX || minX > this.maxAllowedX) {
 			return true;
 		}
