+++ src/sodiumCompatibility/java/net/coderbot/iris/compat/sodium/impl/shadow_map/IrisFrustum.java
@@ -1,5 +1,5 @@
 package net.coderbot.iris.compat.sodium.impl.shadow_map;
 
 public interface IrisFrustum {
-	boolean apply(float minX, float minY, float minZ, float maxX, float maxY, float maxZ);
+	boolean apply(double minX, double minY, double minZ, double maxX, double maxY, double maxZ);
 }
